Absolute base-10 logarithm of p-values

Absolute base-10 logarithm of p-values

absLog10p(x)

Arguments

x
Numeric vector or matrix

The function returns the absolute values of base-10 logarithm of p-values.

Value

Numeric vector or matrix.

Details

The logarithm transformation of p-values is commonly used to visualize results from statistical tests. Although it may cause misunderstanding and therefore its use is disapproved by some experts, it helps to visualize and interpret results of statistical tests intuitively.

The function transforms p-values with base-10 logarithm, and returns its absolute value. The choice of base 10 is driven by the simplicity of interpreting the results.

Author

Jitao David Zhang <jitao_david.zhang@roche.com>

Examples

testp <- runif(1000, 0, 1)
testp.al <- absLog10p(testp)

print(head(testp))

#> [1] 0.080750138 0.834333037 0.600760886 0.157208442 0.007399441 0.466393497

print(head(testp.al))

#> [1] 1.09285673 0.07866056 0.22129835 0.80352414 2.13080108 0.33124751