Shannon entropy

Shannon entropy

entropy(vector)

Arguments

vector
A vector of numbers, or characters. Discrete probability of each item is calculated and the Shannon entropy is returned.

Value

Shannon entropy

Shannon entropy can be used as measures of gene expression specificity, as well as measures of tissue diversity and specialization. See references below.

We use 2 as base for the entropy calculation, because in this base the unit of entropy is bit.

References

Martinez and Reyes-Valdes (2008) Defining diversity, specialization, and gene specificity in transcriptomes through information theory. PNAS 105(28):9709--9714

Author

Jitao David Zhang <jitao_david.zhang@roche.com>

Examples

myVec0 <- 1:9
entropy(myVec0) ## log2(9)

#> [1] 3.169925

myVec1 <- rep(1, 9)
entropy(myVec1)

#> [1] 0

entropy(LETTERS)

#> [1] 4.70044

entropy(rep(LETTERS, 5))

#> [1] 4.70044